fix functions from different versions
