+++ core/src/main/spark_3.0_mount/scala/doric/syntax/All.scala
@@ -22,3 +22,4 @@ private[doric] trait All
     with CommonColumns3x
     with MapColumns3x
     with StringColumn3x
+    with NumericColumns2_31
