+++ core/src/main/spark_3.1_mount/scala/doric/syntax/All.scala
@@ -27,3 +27,4 @@ private[doric] trait All
     with CommonColumns3x
     with MapColumns3x
     with StringColumn3x
+    with NumericColumns2_31
