+++ core/src/main/spark_3.2_mount/scala/doric/syntax/All.scala
@@ -26,3 +26,4 @@ private[doric] trait All
     with CommonColumns3x
     with MapColumns3x
     with StringColumn3x
+    with NumericColumns32
